+++ composeunstyled-modal/src/nonAndroidMain/kotlin/com/composeunstyled/Modal.nonAndroid.kt
@@ -39,7 +39,7 @@ import androidx.compose.ui.window.DialogProperties
 actual fun Modal(
   state: ModalState,
   onKeyEvent: (KeyEvent) -> Boolean,
-  content: @Composable () -> Unit,
+  content: @Composable ModalScope.() -> Unit,
 ) {
   if (
     state.transitionState.targetState.not() &&
@@ -68,7 +68,7 @@ actual fun Modal(
               state.attachedToWindow = false
             }
           }
-          content()
+          ModalScopeInstance.content()
         }
       }
     },