+++ demo/src/commonMain/kotlin/HazeScrimDemo.kt
@@ -59,9 +59,9 @@ import androidx.compose.ui.text.TextStyle
 import androidx.compose.ui.unit.dp
 import com.composables.uripainter.rememberUriPainter
 import com.composeunstyled.DialogPanel
+import com.composeunstyled.Scrim
 import com.composeunstyled.UnstyledButton
 import com.composeunstyled.UnstyledDialog
-import com.composeunstyled.UnstyledScrim
 import dev.chrisbanes.haze.blur.blurEffect
 import dev.chrisbanes.haze.hazeEffect
 import dev.chrisbanes.haze.hazeSource
@@ -100,18 +100,19 @@ fun HazeScrimDemo() {
     UnstyledDialog(
       visible = dialogVisible,
       onDismissRequest = { dialogVisible = false },
+      overlay = {
+        Scrim(
+          modifier = Modifier.hazeEffect(hazeState) {
+            blurEffect {
+              blurRadius = 22.dp
+            }
+          },
+          scrimColor = Color.Transparent,
+          enter = fadeIn(),
+          exit = fadeOut(),
+        )
+      },
     ) {
-      UnstyledScrim(
-        modifier = Modifier.hazeEffect(hazeState) {
-          blurEffect {
-            blurRadius = 22.dp
-          }
-        },
-        scrimColor = Color.Transparent,
-        enter = fadeIn(),
-        exit = fadeOut(),
-      )
-
       DialogPanel(
         modifier = Modifier
           .padding(20.dp)