+++ demo/src/commonMain/kotlin/ModalBottomSheetDemo.kt
@@ -53,6 +53,7 @@ import androidx.compose.ui.draw.shadow
 import androidx.compose.ui.graphics.Brush
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.unit.dp
+import com.composeunstyled.Scrim
 import com.composeunstyled.Sheet
 import com.composeunstyled.SheetDetent
 import com.composeunstyled.SheetDetent.Companion.FullyExpanded
@@ -60,7 +61,6 @@ import com.composeunstyled.SheetDetent.Companion.Hidden
 import com.composeunstyled.UnstyledButton
 import com.composeunstyled.UnstyledDragIndication
 import com.composeunstyled.UnstyledModalBottomSheet
-import com.composeunstyled.UnstyledScrim
 import com.composeunstyled.currentWindowContainerSize
 import com.composeunstyled.focusRing
 import com.composeunstyled.rememberModalBottomSheetState
@@ -105,7 +105,7 @@ fun ModalBottomSheetDemo() {
     UnstyledModalBottomSheet(
       state = modalSheetState,
       overlay = {
-        UnstyledScrim(scrimColor = Color.Black.copy(0.3f), enter = fadeIn(), exit = fadeOut())
+        Scrim(scrimColor = Color.Black.copy(0.3f), enter = fadeIn(), exit = fadeOut())
       },
     ) {
       Sheet(